fix: change styles for correct displaying of page elements
