+++ src/components/NavBar/NavBar.tsx
@@ -16,12 +16,12 @@ const NavBar = () =>  {
     return (
         <div>
             <Box sx={{ flexGrow: 1 }}>
-                <AppBar className='navbar' position='absolute'>
+                <AppBar className='navbar'>
                     <Toolbar className='toolbar' >
-                        <NavBarIcon active={iconActive} setActive={setIconActive}/>
-                        <SelectBuilding/>
-                        <NavBarSwitch active={iconActive} setActive={setIconActive}/>
-                        <NavBarInformation/>
+                        <NavBarIcon active={iconActive} setActive={setIconActive} />
+                        <SelectBuilding />
+                        <NavBarSwitch active={iconActive} setActive={setIconActive} />
+                        <NavBarInformation />
                     </Toolbar>
                 </AppBar>
             </Box>
